+++ src/main/java/com/project/orderinventorymanagement/customerservice/repository/CustomerRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.orderinventorymanagement.customerservice.entity.Customer;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 }
